# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Secure Online Election Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── USERS (Admin) ─────────────────────────────────────────────────────────
...
  /elections/{electionId}/candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates for an election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Add candidate to election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /elections/{electionId}/candidates/{candidateId}:
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Remove candidate
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── VOTERS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── AUTH ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [fullName, email, password, role]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [voter, election_creator]
        orgName:
          type: ["string", "null"]
        electionPurpose:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ─── USER ────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, fullName, email, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, election_creator, voter]
        status:
          type: string
          enum: [active, blocked, pending]
        orgName:
          type: ["string", "null"]
        electionPurpose:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Candidate:
      type: object
      required: [id, electionId, name, createdAt]
      properties:
        id:
          type: integer
        electionId:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        manifesto:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        voteCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    CandidateInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        manifesto:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        manifesto:
          type: ["string", "null"]
        description:
          type: ["string", "null"]

    # ─── VOTER ───────────────────────────────────────────────────────────────
...