# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Secure Online Election Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  orgName?: string | null;
  /** @nullable */
  electionPurpose?: string | null;
  createdAt: Date;
}